refactor: move hybrid proxy handlers into server behind a proxy feature
Relocate the per-surface upstream proxy adapters from the cli crate into a
reusable unitycatalog-server::handlers module so the reference implementation
(and future patterns like federation) can use them, gated behind a new optional
`proxy` feature that pulls unitycatalog-client only when enabled.

- server: add optional unitycatalog-client dep + proxy/federation features (not
  default); new handlers module with handlers::upstream (gated) holding the
  Upstream{Catalog,Schema,Table}Handler adapters + map_upstream_err, all pub,
  generic over Cx, with no RequestContext reference so a later crate split stays
  mechanical

- server::policy: add dyn-friendly filter_authorized(&dyn Policy<Cx>, ...) next
  to process_resources (which now delegates to it), removing the Arc<dyn Policy>
  Sized workaround the proxy needed

- cli: hybrid.rs slimmed to run_server_rest_hybrid wiring only, importing the
  adapters from unitycatalog_server::handlers::upstream; server dep now enables
  ["axum", "proxy"]

Verified: server builds without proxy and does not pull the path client
(cargo tree); --features proxy does; clippy clean; config tests pass.
