+++ crates/server/Cargo.toml
@@ -14,6 +14,7 @@ unitycatalog-common = { path = "../common", features = ["axum"] }
 unitycatalog-sharing-client = { path = "../sharing-client", features = [
   "server",
 ] }
+unitycatalog-client = { path = "../client", optional = true }
 
 async-trait = { workspace = true }
 bytes = { workspace = true }
@@ -60,3 +61,7 @@ axum = ["dep:axum", "dep:axum-extra", "http", "tower"]
 memory = ["dashmap", "uuid"]
 
 grpc = ["tonic"]
+
+# reusable handler patterns: proxy leaves + UC-client connector
+proxy = ["dep:unitycatalog-client"]
+federation = ["proxy"]
